Rename `StackManager` to `LifecycleManager`.
