+++ packages/breadboard/src/index.ts
@@ -13,7 +13,7 @@ export { SchemaBuilder, combineSchemas } from "./schema.js";
 export { RunResult } from "./run.js";
 export { TraversalMachine } from "./traversal/machine.js";
 export { MachineResult } from "./traversal/result.js";
-export { traversalResultFromStack } from "./stack.js";
+export { traversalResultFromStack } from "./run/stack.js";
 export { toMermaid } from "./mermaid.js";
 export { callHandler } from "./handler.js";
 export { asRuntimeKit } from "./kits/ctors.js";
