+++ packages/breadboard/src/run/index.ts
@@ -4,11 +4,11 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { StackManager } from "../stack.js";
+import { LifecycleManager } from "./stack.js";
 import { ManagedRunState } from "../types.js";
 
 export const createRunStateManager = (): ManagedRunState => {
-  const lifecycle = new StackManager();
+  const lifecycle = new LifecycleManager();
   return {
     lifecycle() {
       return lifecycle;
